Add image rpi-serial-monitor
Boot this image on a raspberry pi board to
monitor the serial line of another board.

Kernel arguments console=.* are discarded by u-boot scripts
on this image, to avoid having this monitoring board
send characters to the other board.
However, when rebooting, u-boot will still send chars.

Note that if you are using 3B, 3B+ or 4B boards,
you should uncomment the following line in config.txt,
on both boards (the one monitoring and the one monitored),
in order to use the more capable PL011 chip instead of
the mini-UART:
dtoverlay=disable-bt
(but this will disable bluetooth)