Enable walt image shell advanced mode on debian images
Provide /bin/walt-image-shell-start and /bin/walt-image-shell-shutdown
scripts to start the OS init (systemd) when running "walt image shell"
and shut it down on exit.
Disallow startup of some systemd services which do not make sense
in "walt image shell", by using "ConditionVirtualization=false"
(e.g., in rpi-sd-update.service) or "ConditionVirtualization=!container"
when it still makes sense to run it in a virtual node (e.g., in
walt-logs.service).
Since walt-logs.service is embedded in the walt-node python wheel,
this wheel and its dependency walt-common were updated too.
When trying to remove services which do not work well in "walt image shell"
or do not make sense in this environment, we tried to avoid adding code
to /bin/walt-image-shell-start and /bin/walt-image-shell-shutdown, and
instead use "ConditionVirtualization=!container" settings for instance;
because file changes done once in the container are saved in the overlay
filesystem layer and accumulate in the image at each shell session.
The only spurious changes we could not prevent are the `/etc/machine-id`
file, which systemd seems to rewrite each time even if it does not change,
and `/var/tmp` whose date of last modification is updated. Other file
changes are those really done in the shell session.
While tracking those spurious changes, we noted that the OS also had
a few systemd `timer` units configured. Those are maintenance tasks
that should not occur during an experiment, so we disabled them too.