rpi64 image, 1-uboot-transfer mode
Provide a new rpi64 image compatible with rpi nodes >= 3B.
Set it as default starting from 3B+.

Handle 1-uboot-transfer mode on rpi32 and rpi64 images, where u-boot
tries downloading fit-start.uboot first, then start.uboot as a fallback
(see walt-node-boot repo). fit-start.uboot is a FIT image containing
the startup script (previously the content of start.uboot).

The nbfs feature handling was previously relying on various variants
of start.uboot for protocols nfs, nfs4 and nbfs, and start.uboot was
actually a symlink to the selected variant. The script activate-nbfs.sh
referenced in file image.spec was updating this symlink when the nbfs
feature was found both in the image and on the server. Since u-boot
may now just download a single FIT image containing the boot script,
it was not reasonable to do the same because it would imply to have
three versions of the FIT image. Instead, we decided to no longer
support NBFS on 32bit rpi images (rpi32-alpine, rpi32-debian) and
on nanopi-r5c. Only PC images (booting with PXE, not u-boot) and
the new rpi64 image support it now. The rpi64 image supports it
by providing several variants of the cmdline.txt file, each of these
variants defining a different u-boot:fs_proto=[nfs|nfs4|nbfs] option.
The rpi32 image cannot work this way because older models have no
network-boot capable firmware, so cmdline.txt is on the SD card, not
inside the image.

The nbfs static binaries are now provided in separate overlay
subdirectories.