rpi64-debian: fix updating initramfs on node (/boot/firmware)
Raspberry Pi OS expects /boot/firmware to be a mount-point
(on SD card), and a post-update script of update-initramfs
checks that.

So this commit adds the definition of a bind-mount of
/boot/firmware on itself, in file /etc/systemd/system/boot-firmware.mount.

The file also has a condition ConditionVirtualization=!container
in order to avoid trying this mount in walt image shell (because
mount operations fail there).
This does not prevent update-initramfs from working in walt image shell
because the post-update script of update-initramfs has an internal
check for virtualization, in which case the check for /boot/firmware
to be a mount-point is bypassed.

Note: since we needed this parameter ConditionVirtualization=!container,
we could *not* implement the bind-mount by just adding a line in fstab.