rpi: simplify overlays for boot files
Previously for an image such as rpi32-debian, the symlinks
in /boot/rpi-*/ directories were coming from 2 different
overlays:
* overlays/rpi32/boot/rpi-*/
* overlays/rpi32-debian/boot/rpi-*/

Same for rpi32-alpine and rpi64-debian.

They are now all in overlays/rpi32-debian/ (and
overlays/rpi32-alpine/, overlays/rpi64-debian respectively).

This meant duplicating the ones coming from overlays/rpi32
in overlays/rpi32-debian and overlays/rpi32-alpine, but
since we are just talking of symlinks, this is not a big
deal.

It is simpler having all symlinks in the same overlay
directory, with the name of this directory reflecting
the name of the target image.